+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 
 * Updates
     * (BC Break) google url default domain is now ``"www.google.com`` instead of ``google.com``. This way we avoid extra redirections too ``"www"`` subdomain.
+    * Fix parsing for classical results on mobiles.
 
 ## 0.2.5